+++ lib/Source.mli
@@ -81,3 +81,5 @@ val locs_of_interval : t -> Location.t -> Location.t * Location.t
 val loc_of_pat_constant : t -> Parsetree.pattern -> Location.t
 
 val loc_of_expr_constant : t -> Parsetree.expression -> Location.t
+
+val is_quoted_string : t -> Location.t -> bool